Prevent migration BigbluebuttonRailsTo200E from breaking
When the index doesn't exist.

+++ db/migrate/20141210192333_bigbluebutton_rails_to200_e.rb
@@ -1,6 +1,6 @@
 class BigbluebuttonRailsTo200E < ActiveRecord::Migration
   def up
-    remove_index :bigbluebutton_rooms, :voice_bridge
+    remove_index :bigbluebutton_rooms, :voice_bridge if index_exists?(:books, :created_at)
   end
 
   def down
